+++ e2e/tests/security.spec.ts
@@ -1,33 +1,32 @@
-import { expect, test } from "@playwright/test";
+import { expect } from "@playwright/test";
 
-import { goToHome } from "./utils/helpers";
-import { mockAuthAPI, mockConfigAPI, mockTidalQueries } from "./utils/mock";
+import { test } from "../test-isolation";
 
+test.use({ envFile: ".env.e2e.auth" });
 test("Tidarr security : Should see login page", async ({ page }) => {
-  await mockAuthAPI(page, "tokenABCXYZ");
-  await mockConfigAPI(page);
-  await goToHome(page);
+  await page.goto("/");
   await page.evaluate("localStorage.clear()");
   await expect(
     page.getByRole("heading", { name: "Tidarr authentication" }),
   ).toBeInViewport();
 
   // When I proceed to login
-  await page.getByPlaceholder("Password...").fill("tidarrpwd");
+  await page.getByPlaceholder("Password...").fill("passwdtidarr");
   await page.getByRole("button", { name: "Submit" }).click();
 
   // Then I should be on homepage
   await expect(page.getByRole("heading", { name: "Tidarr" })).toBeInViewport();
+  await page.waitForTimeout(500);
 
   // And A new token should be set
   const token = await page.evaluate("localStorage.getItem('tidarr-token')");
-  await expect(token).toEqual("tokenABCXYZ");
+  await expect(token).not.toBeNull();
 
   // When I remove token and navigate to a different page to trigger re-auth check
   await page.evaluate("localStorage.removeItem('tidarr-token')");
 
   // When I go to homepage
-  await goToHome(page);
+  await page.goto("/");
   // Then I should be redirected to login page
   await expect(
     page.getByRole("heading", { name: "Tidarr authentication" }),
@@ -37,12 +36,11 @@ test("Tidarr security : Should see login page", async ({ page }) => {
 test("Tidarr security : Should be redirected to login page", async ({
   page,
 }) => {
-  await mockAuthAPI(page, "tokenABCXYZ");
   await page.goto("/login");
   await page.evaluate("localStorage.clear()");
 
   // When I go to homepage
-  await goToHome(page);
+  await page.goto("/");
   // Then I should be redirected to login page
   await expect(
     page.getByRole("heading", { name: "Tidarr authentication" }),
@@ -52,16 +50,13 @@ test("Tidarr security : Should be redirected to login page", async ({
 test("Tidarr security : Should be redirected to requested url after login", async ({
   page,
 }) => {
-  await mockAuthAPI(page, "tokenABCXYZ");
-  await mockConfigAPI(page);
-  await mockTidalQueries(page);
   await page.goto("/search/Nirvana");
   await expect(
     page.getByRole("heading", { name: "Tidarr authentication" }),
   ).toBeInViewport();
 
   // When I proceed to login
-  await page.getByPlaceholder("Password...").fill("tidarrpwd");
+  await page.getByPlaceholder("Password...").fill("passwdtidarr");
   await page.getByRole("button", { name: "Submit" }).click();
 
   // Then I should be on homepage
@@ -71,14 +66,11 @@ test("Tidarr security : Should be redirected to requested url after login", asyn
 });
 
 test("Tidarr security : Should be able to log out", async ({ page }) => {
-  await mockAuthAPI(page, "tokenABCXYZ");
-  await mockConfigAPI(page);
-  await mockTidalQueries(page);
   await page.goto("/login");
   await page.evaluate("localStorage.clear()");
 
   // When I go to login page and log in
-  await page.getByPlaceholder("Password...").fill("tidarrpwd");
+  await page.getByPlaceholder("Password...").fill("passwdtidarr");
   await page.getByRole("button", { name: "Submit" }).click();
 
   // Then I should be on the homepage
@@ -98,9 +90,6 @@ test("Tidarr security : Login page should redirect to home in public mode", asyn
   page,
 }) => {
   // When I go to login page in public mode
-  await mockTidalQueries(page);
-  await mockConfigAPI(page);
-
   await page.goto("/login");
   await page.evaluate("localStorage.clear()");
   // then I should be redirected to homepage
@@ -110,20 +99,19 @@ test("Tidarr security : Login page should redirect to home in public mode", asyn
 test("Tidarr security : Wrong credentials should display an error", async ({
   page,
 }) => {
-  await mockAuthAPI(page, "tokenABCXYZ");
   await page.route("*/**/auth", async (route) => {
     await route.fulfill({
       status: 401,
       json: { error: true, message: "Invalid credentials" },
     });
   });
-  await goToHome(page);
+  await page.goto("/");
   await expect(
     page.getByRole("heading", { name: "Tidarr authentication" }),
   ).toBeInViewport();
 
   // When I proceed to login
-  await page.getByPlaceholder("Password...").fill("tidarrpwd");
+  await page.getByPlaceholder("Password...").fill("passwdtidarr");
   await page.getByRole("button", { name: "Submit" }).click();
 
   // Then I should see error message