+++ e2e/tests/utils/helpers.ts
@@ -1,7 +1,5 @@
 import { expect, Page } from "@playwright/test";
 
-import { mockTidalQueries } from "./mock";
-
 export async function waitForLoader(page: Page) {
   const loader = await page.getByTestId("loader")?.isVisible();
   if (loader) {
@@ -12,7 +10,7 @@ export async function waitForLoader(page: Page) {
 export async function waitForImgLoaded(page: Page) {
   await page.waitForFunction(() => {
     const images = Array.from(document.querySelectorAll("img"));
-    return images.every((img) => img.complete);
+    return images.every((img) => img?.complete);
   });
 }
 
@@ -62,43 +60,4 @@ export async function testProcessingList(
   // await page.getByRole("button", { name: "close" }).first().click();
   await page.locator(".MuiDialog-container button").first().click();
   await page.waitForTimeout(250);
-}
-
-export async function emptyProcessingList(page: Page) {
-  const isVisible = await page.locator("button.MuiFab-circular").isVisible();
-
-  if (!isVisible) return null;
-
-  // Open the processing list if not already open
-  const isExpanded = await page
-    .locator("button.MuiFab-circular")
-    .getAttribute("aria-expanded");
-  if (isExpanded !== "true") {
-    await page.locator("button.MuiFab-circular").click();
-    // Wait for the table to be visible
-    await page.waitForSelector('[aria-label="Processing table"]', {
-      state: "visible",
-      timeout: 5000,
-    });
-  }
-
-  await page.waitForTimeout(250);
-  const clearButton = await page.getByRole("button", {
-    name: "Clear finished",
-  });
-  const isClearButtonVisible = await clearButton.isVisible();
-  const isPageClosed = page.isClosed();
-  if (!isPageClosed && isClearButtonVisible) await clearButton.click();
-  await page.waitForTimeout(250);
-}
-
-export async function emptySyncList(page: Page) {
-  // Call the API to clear all sync items
-  await page.request.delete("http://localhost:8484/api/sync/remove-all");
-  await page.waitForTimeout(250);
-}
-
-export async function goToHome(page: Page) {
-  await mockTidalQueries(page);
-  await page.goto("/");
 }